+++ spec/unit/resource/api800/synergy/hypervisor_cluster_profile_spec.rb
@@ -1,0 +1,9 @@
+require 'spec_helper'
+
+RSpec.describe OneviewSDK::API800::Synergy::HypervisorClusterProfile do
+  include_context 'shared context'
+
+  it 'inherits from OneviewSDK::API800::C7000::HypervisorClusterProfile' do
+    expect(described_class).to be < OneviewSDK::API800::C7000::HypervisorClusterProfile
+  end
+end